Custom Schedules
added custom schedules for
daily
twice daily
hourly

to work as expected by user.

by defaul wp only allows to change the initial run of the cron, later runs for daily and twicedayily will occur with said intervals, but at convenient for wp times.

this means that if a cron is set to initially run at 16:00 daily, the first run will occure at 16:00, but later runs will not have to occur at that time.